% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autismSexDifferences.R
\docType{package}
\name{autismSexDifferences}
\alias{autismSexDifferences}
\alias{autismSexDifferences-package}
\title{autismSexDifferences: An R package for looking at ASD effects on neuroanatomy modulated by sex}
\description{
Code to run Chris Hammill's OHBM analysis examining how sex
interacts and modulates the effect of autism spectrum disorders
on neuroanatomy
}
