% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{evidence_ratio}
\alias{evidence_ratio}
\title{Compute evidence ratios}
\usage{
evidence_ratio(aics)
}
\arguments{
\item{aics}{Vector of AIC values}
}
\value{
A vector of evidence ratios
}
\description{
Compute the ratio of evidence for the best model relative to each competing model
}
