% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{fit_models}
\alias{fit_models}
\title{Fit a common suite of anatLms for model selection}
\usage{
fit_models(form_list = c(nint = ~sex + Dx + bv + age_scan + Dx:sex +
  new_scanner, nint_crb = ~sex + Dx + crbv + age_scan + Dx:sex + new_scanner,
  nint_ttrb = ~sex + Dx + ttrbv + age_scan + Dx:sex + new_scanner, pnint = ~sex
  + Dx + bv + poly(age_scan, 2) + Dx:sex + new_scanner, pnint_crb = ~sex + Dx +
  crbv + poly(age_scan, 2) + Dx:sex + new_scanner, pnint_ttrb = ~sex + Dx +
  ttrbv + poly(age_scan, 2) + Dx:sex + new_scanner, nex = ~sex + Dx + bv +
  age_scan + Dx:sex + age_scan:sex + new_scanner, nex_crb = ~sex + Dx + crbv +  
     age_scan + Dx:sex + age_scan:sex + new_scanner, nex_ttrb = ~sex + Dx +
  ttrbv + age_scan + Dx:sex + age_scan:sex + new_scanner, nage = ~sex + Dx + bv
  + age_scan + Dx:sex + bv:sex + new_scanner, nage_crb = ~sex + Dx + crbv +
  age_scan + Dx:sex + crbv:sex + new_scanner, nage_ttrb = ~sex + Dx + ttrbv +
  age_scan + Dx:sex + ttrbv:sex + new_scanner, st = ~sex + Dx + bv + age_scan +
  Dx:sex + bv:sex + age_scan:sex + new_scanner, crb = ~sex + Dx + crbv +
  age_scan + Dx:sex + crbv:sex + age_scan:sex + new_scanner,      ttrb = ~sex +
  Dx + ttrbv + age_scan + Dx:sex + ttrbv:sex + age_scan:sex + new_scanner, page
  = ~sex + Dx + bv + poly(age_scan, 2) + Dx:sex + bv:sex + poly(age_scan, 2):sex
  + new_scanner, page_crb = ~sex + Dx + crbv + poly(age_scan, 2) + Dx:sex +
  crbv:sex + poly(age_scan, 2):sex + new_scanner, page_ttrb = ~sex + Dx + ttrbv
  + poly(age_scan, 2) + Dx:sex + ttrbv:sex + poly(age_scan, 2):sex +
  new_scanner), anat, model_data, ...)
}
\arguments{
\item{form_list}{A named list of one sided formulae for model selection}

\item{anat}{A matrix of anatomy values}

\item{model_data}{A data frame of covariates}
}
\value{
A list of \code{anatModel} objects for model selection
}
\description{
Fit all models to a common anat object and covariate data.frame
for use in model selection
}
