% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{full_match}
\alias{full_match}
\title{Perform full matching}
\usage{
full_match(data, dist_mat, formula = Dx ~ age_scan + bv + iq + sex, ...)
}
\arguments{
\item{data}{The data.frame of interest}

\item{dist_mat}{The n x n distance matrix}

\item{formula}{The formula for the matching}

\item{...}{Additional arguments to matchit and optmatch below.}
}
\value{
data augmented with \code{weight} and
\code{subclass} columns
}
\description{
Given a formula, a distance matrix and a data.frame
perform full matching on the rows of data. Returns
the data frame augmented with \code{weight} and
\code{subclass} columns
}
