% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{mids_to_dfs}
\alias{mids_to_dfs}
\title{Convert mids to list of data frames}
\usage{
mids_to_dfs(mids)
}
\arguments{
\item{mids}{The multiply imputed data set from \link[mice]{mice}}
}
\value{
A list of \code{data.frames}
}
\description{
Utility function extract a \code{mids} object into a
list of \code{data.frames}
}
