% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{rank_AICs}
\alias{rank_AICs}
\title{Rank AICs}
\usage{
rank_AICs(aics, form_list)
}
\arguments{
\item{form_list}{A named list of formulae, extracted from the anatLm list if not supplied and
possible}
}
\value{
A \code{tibble} with the model names, formulae, and evidence ratios
}
\description{
Take a set of AICS and produce a nice ranking tibble
}
