% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{rank_models}
\alias{rank_models}
\title{Model ranking}
\usage{
rank_models(model_list, form_list = NULL)
}
\arguments{
\item{model_list}{A list of anatLm models}

\item{form_list}{A named list of formulae, extracted from the anatLm list if not supplied and
possible}
}
\value{
A \code{tibble} with the model names, formulae, and evidence ratios
}
\description{
Take a list of anatLms and rank the models based on minimum median evidence ratio
}
