% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching.R
\name{subj_dist}
\alias{subj_dist}
\title{Custom matching distance function}
\usage{
subj_dist(i, j, data, w_iq = 1/2, w_bv = 1/4)
}
\arguments{
\item{data}{A data frame containing Dx, sex, bv, and
iq columns}

\item{w_iq}{The weight for iq (as a fraction of the weight for age}

\item{w_bv}{The weight for brain-volume as a fraction of the weight
for age}
}
\value{
The distance between subjects x and y
}
\description{
Used for the OHBM analysis, exact matching on sex
euclidean distance for age after scaling, half
weight for IQ, quarter weight for brain volume by
default.
}
